+++ src/sample/cnv_parts.py
@@ -4,7 +4,7 @@ Copyright (c) 2021 Peter Triesberger
 For further information see https://github.com/peter88213/aeon3odt
 Published under the MIT License (https://opensource.org/licenses/mit-license.php)
 """
-SUFFIX = '_parts'
+SUFFIX = '_chapter_overview'
 
 SETTINGS = dict(
     part_number_prefix='Part',